requirements.txt
pylint

numpy
pillow
matplotlib
